math_adapter : fix several problems, clean tests
