+++ module/rust/math_adapter/rust/impl/math/plugin/cgmath/x2.rs
@@ -5,6 +5,8 @@ pub mod internal
   use crate::prelude::*;
   use crate::{ X2, ScalarInterface };
 
+  //
+
   macro_rules! impl_x2_for
   {
 
@@ -158,9 +160,8 @@ pub mod internal
 
   }
 
-  impl_x2_for!( cgmath::Vector2< Scalar > );
   impl_x2_as_for!( cgmath::Vector2< Scalar > );
-
+  impl_x2_for!( cgmath::Vector2< Scalar > );
   impl_x2_for!( cgmath::Point2< Scalar > );
 
 }
