+++ module/rust/math_adapter/rust/impl/math/scalar.rs
@@ -1,7 +1,7 @@
 /// Internal namespace.
 pub mod internal
 {
-  use core::fmt::Debug;
+  use core::fmt::{ Debug, Display };
   // use core::hash::Hash;
   use core::ops::{ Add, Sub, Mul, Div, Rem, AddAssign, SubAssign, MulAssign, DivAssign, RemAssign };
   use std::cmp::PartialOrd;
@@ -14,6 +14,7 @@ pub mod internal
   pub trait ScalarInterface :
     'static +
     Debug +
+    Display +
     PartialEq +
     PartialOrd +
     Copy +
@@ -42,6 +43,7 @@ pub mod internal
     T :
       'static +
       Debug +
+      Display +
       PartialEq +
       PartialOrd +
       Copy +
