+++ module/rust/math_adapter/rust/test/math/cgmath_test/mod.rs
@@ -1,5 +1,15 @@
 
 mod x2_test;
-mod as_test;
+
+
+#[
+  cfg( all
+  (
+    not( feature = "nalgebra_ops" ),
+    not( all( feature = "default_ops", feature = "nalgebra" ) ),
+    any( feature = "default_ops", feature = "cgmath_ops" ),
+  ))
+]
+mod ops_test;
 
 /* xxx : merge in */
