+++ module/rust/math_adapter/rust/test/math/cgmath_test/x2_test.rs
@@ -6,13 +6,13 @@ use num_traits::cast::cast;
 use math_adapter::prelude::*;
 use math_adapter::X2;
 use crate::test_tools::*;
-use crate::{ macro_test_x2_with_records, num };
+use crate::{ num };
 
 ///
 /// One test should be ordinary to exclude possibility of problems with macro.
 ///
 
-fn _basic()
+fn _basic_test()
 {
   type T = i8;
 
@@ -41,17 +41,19 @@ fn _basic()
 
 }
 
-//
+///
+/// Parametrized test.
+///
 
 fn _canonical_test()
 {
-  macro_test_x2_with_records!( cgmath::Vector2, x, y ; i8, i16, i32, i64, i128 );
-  macro_test_x2_with_records!( cgmath::Vector2, x, y ; u8, u16, u32, u64, u128 );
-  macro_test_x2_with_records!( cgmath::Vector2, x, y ; f32, f64 );
+  crate::macro_x2::macro_test_x2_with_records!( cgmath::Vector2, x, y ; i8, i16, i32, i64, i128 );
+  crate::macro_x2::macro_test_x2_with_records!( cgmath::Vector2, x, y ; u8, u16, u32, u64, u128 );
+  crate::macro_x2::macro_test_x2_with_records!( cgmath::Vector2, x, y ; f32, f64 );
 
-  macro_test_x2_with_records!( cgmath::Point2, x, y ; i8, i16, i32, i64, i128 );
-  macro_test_x2_with_records!( cgmath::Point2, x, y ; u8, u16, u32, u64, u128 );
-  macro_test_x2_with_records!( cgmath::Point2, x, y ; f32, f64 );
+  crate::macro_x2::macro_test_x2_with_records!( cgmath::Point2, x, y ; i8, i16, i32, i64, i128 );
+  crate::macro_x2::macro_test_x2_with_records!( cgmath::Point2, x, y ; u8, u16, u32, u64, u128 );
+  crate::macro_x2::macro_test_x2_with_records!( cgmath::Point2, x, y ; f32, f64 );
 
   // trace_macros!( true );
   // macro_test_x2_with_records!( cgmath::Point2, x, y ; f32 );
@@ -59,10 +61,23 @@ fn _canonical_test()
 
 }
 
+///
+/// Tests for X2 conversion function. Names are implementation-specific. .
+///
+
+fn _convertion_as_specific_test()
+{
+  type T = i8;
+
+  crate::macro_x2::macro_test_x2_as_specific!( cgmath::Vector2, cgmath ; T );
+
+}
+
 //
 
 test_suite!
 {
-  basic,
+  basic_test,
   canonical_test,
+  convertion_as_specific_test,
 }
