+++ module/rust/math_adapter/rust/test/math/nalgebra_test/x2_test.rs
@@ -6,13 +6,13 @@ use num_traits::cast::cast;
 use math_adapter::prelude::*;
 use math_adapter::X2;
 use crate::test_tools::*;
-use crate::{ macro_test_x2_with_records, num };
+use crate::{ num };
 
 ///
 /// One test should be ordinary to exclude possibility of problems with macro.
 ///
 
-fn _basic()
+fn _basic_test()
 {
   type T = i8;
 
@@ -48,11 +48,76 @@ fn _basic()
 
 //
 
+// macro_rules! callback
+// {
+//   ( $callback:ident( $( $args:tt )* ) ) =>
+//   {
+//     $callback!( $( $args )* )
+//   };
+// }
+
+//
+//
+// macro_rules! primitives_numbers
+// {
+//   () =>
+//   {
+//     i8, i16, i32, i64, i128, u8, u16, u32, u64, u128, f32, f64,
+//   }
+// }
+//
+// //
+//
+// macro_rules! generate_fn
+// {
+//   ( $Type : ident ) =>
+//   {
+//     fn inc( src : $Type )
+//     {
+//       src + 1;
+//     }
+//   }
+// }
+
+//
+
+// #[ macro_export ]
+// macro_rules! apply
+// {
+//   ( $Callback:ident ( $( $Prefix:tt )* ) @Args $( $Arg:tt ),* ) =>
+//   {
+//     $( $Callback!( $( $Prefix )* $Arg ); )*
+//   }
+// }
+//
+// //
+//
+// #[ macro_export ]
+// macro_rules! for_each_number
+// {
+//   ( $Callback:ident $( $Prefix:tt )* ) =>
+//   {
+//     apply!( $Callback ( $( $Prefix )* ) @Args i8, i16, i32, i64, i128, u8, u16, u32, u64, u128, f32, f64 )
+//   }
+// }
+
+///
+/// Parametrized test.
+///
+
 fn _canonical_test()
 {
-  macro_test_x2_with_records!( nalgebra::Vector2, x, y ; i8, i16, i32, i64, i128 );
-  macro_test_x2_with_records!( nalgebra::Vector2, x, y ; u8, u16, u32, u64, u128 );
-  macro_test_x2_with_records!( nalgebra::Vector2, x, y ; f32, f64 );
+
+  // for_each_number!( macro_test_x2_with_records, nalgebra::Vector2, x, y ; );
+
+  crate::macro_x2::macro_test_x2_with_records!( nalgebra::Vector2, x, y ; i8, i16, i32, i64, i128 );
+  crate::macro_x2::macro_test_x2_with_records!( nalgebra::Vector2, x, y ; u8, u16, u32, u64, u128 );
+  crate::macro_x2::macro_test_x2_with_records!( nalgebra::Vector2, x, y ; f32, f64 );
+
+  crate::macro_x2::macro_test_x2_with_records!( nalgebra::geometry::Point2, x, y ; i8, i16, i32, i64, i128 );
+  crate::macro_x2::macro_test_x2_with_records!( nalgebra::geometry::Point2, x, y ; u8, u16, u32, u64, u128 );
+  crate::macro_x2::macro_test_x2_with_records!( nalgebra::geometry::Point2, x, y ; f32, f64 );
+
   /* zzz : use callback instead? */
 
   // trace_macros!( true );
@@ -61,10 +126,24 @@ fn _canonical_test()
 
 }
 
+///
+/// Tests for X2 conversion function. Names are implementation-specific. .
+///
+
+#[ test ]
+fn _convertion_as_specific_test()
+{
+  type T = i8;
+
+  crate::macro_x2::macro_test_x2_as_specific!( nalgebra::Vector2, nalgebra ; T );
+
+}
+
 //
 
 test_suite!
 {
-  basic,
+  basic_test,
   canonical_test,
+  convertion_as_specific_test
 }
