+++ module/rust/math_adapter/rust/test/math/winit_test/x2_test.rs
@@ -1,12 +1,11 @@
-#![ allow( unused_imports ) ]
+// #![ allow( unused_imports ) ]
 
 use wtest_basic::*;
 use core::mem::size_of;
 use math_adapter::prelude::*;
 use math_adapter::X2;
-use num_traits::cast::cast;
 use crate::test_tools::*;
-use crate::{ macro_test_x2_with_records, num };
+use crate::{ num };
 
 ///
 /// One test should be ordinary to exclude possibility of problems with macro.
@@ -48,21 +47,62 @@ fn _basic()
 fn _physical_size_test()
 {
 
-  macro_test_x2_with_records!( winit::dpi::PhysicalSize, width, height ; i8, i16, i32, i64, i128 );
-  macro_test_x2_with_records!( winit::dpi::PhysicalSize, width, height ; u8, u16, u32, u64, u128 );
-  macro_test_x2_with_records!( winit::dpi::PhysicalSize, width, height ; f32, f64 );
+  crate::macro_x2::macro_test_x2_with_records!( winit::dpi::PhysicalSize, width, height ; i8, i16, i32, i64, i128 );
+  crate::macro_x2::macro_test_x2_with_records!( winit::dpi::PhysicalSize, width, height ; u8, u16, u32, u64, u128 );
+  crate::macro_x2::macro_test_x2_with_records!( winit::dpi::PhysicalSize, width, height ; f32, f64 );
 
-  macro_test_x2_with_records!( winit::dpi::LogicalSize, width, height ; i8, i16, i32, i64, i128 );
-  macro_test_x2_with_records!( winit::dpi::LogicalSize, width, height ; u8, u16, u32, u64, u128 );
-  macro_test_x2_with_records!( winit::dpi::LogicalSize, width, height ; f32, f64 );
+  crate::macro_x2::macro_test_x2_with_records!( winit::dpi::LogicalSize, width, height ; i8, i16, i32, i64, i128 );
+  crate::macro_x2::macro_test_x2_with_records!( winit::dpi::LogicalSize, width, height ; u8, u16, u32, u64, u128 );
+  crate::macro_x2::macro_test_x2_with_records!( winit::dpi::LogicalSize, width, height ; f32, f64 );
 
-  macro_test_x2_with_records!( winit::dpi::PhysicalPosition, x, y ; i8, i16, i32, i64, i128 );
-  macro_test_x2_with_records!( winit::dpi::PhysicalPosition, x, y ; u8, u16, u32, u64, u128 );
-  macro_test_x2_with_records!( winit::dpi::PhysicalPosition, x, y ; f32, f64 );
+  crate::macro_x2::macro_test_x2_with_records!( winit::dpi::PhysicalPosition, x, y ; i8, i16, i32, i64, i128 );
+  crate::macro_x2::macro_test_x2_with_records!( winit::dpi::PhysicalPosition, x, y ; u8, u16, u32, u64, u128 );
+  crate::macro_x2::macro_test_x2_with_records!( winit::dpi::PhysicalPosition, x, y ; f32, f64 );
 
-  macro_test_x2_with_records!( winit::dpi::LogicalPosition, x, y ; i8, i16, i32, i64, i128 );
-  macro_test_x2_with_records!( winit::dpi::LogicalPosition, x, y ; u8, u16, u32, u64, u128 );
-  macro_test_x2_with_records!( winit::dpi::LogicalPosition, x, y ; f32, f64 );
+  crate::macro_x2::macro_test_x2_with_records!( winit::dpi::LogicalPosition, x, y ; i8, i16, i32, i64, i128 );
+  crate::macro_x2::macro_test_x2_with_records!( winit::dpi::LogicalPosition, x, y ; u8, u16, u32, u64, u128 );
+  crate::macro_x2::macro_test_x2_with_records!( winit::dpi::LogicalPosition, x, y ; f32, f64 );
+
+}
+
+///
+/// Basic test of interoperability of `winit` with `cgmath`.
+///
+
+#[ cfg( all( feature = "winit", feature = "cgmath" ) ) ]
+#[ test ]
+fn cgmath_winit_interoperability_test()
+{
+
+  /* test.case = "use cgmath vectors for operations on winit vectors"; */
+  {
+    let src1 = winit::dpi::PhysicalSize::< i8 >::make( 3, 2 );
+    let src2 = winit::dpi::PhysicalSize::< i8 >::make( 0, 1 );
+    let got = src1.as_cgmath() - src2.as_cgmath();
+    let exp = cgmath::Vector2::< i8 >::make( 3, 1 );
+    assert_eq!( got, exp );
+  }
+
+}
+
+///
+/// Interoperability with winit.
+///
+
+#[ cfg( all( feature = "winit" ) ) ]
+#[ cfg( any( feature = "nalgebra_ops", feature = "default_ops" ) ) ]
+#[ test ]
+fn inter_winit()
+{
+  type T = i8;
+
+  /* test.case = "basic"; */
+  {
+    let src1 = winit::dpi::PhysicalSize::< T >::make( 1, 3 );
+    let got = src1.as_canonical().sum();
+    let exp = 4;
+    assert_eq!( got, exp );
+  }
 
 }
 
@@ -73,3 +113,4 @@ test_suite!
   basic,
   physical_size_test,
 }
+/* xxx : extend the macro */
